requirements.txt
requests
openpyxl
xlrd
six
pyyaml
datapackage
